+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [1.1.8] - (2021-Apr-20)
+
+* Fixed bug that was preventing copying of dark background.  Fixes [#163](https://github.com/jonbhanson/flutter_native_splash/issues/163).
+* Updated pubspec.yaml to show which platforms are supported.
+
 ## [1.1.7+1] - (2021-Apr-02)
 
 * flutter_native_splash:remove adheres to android/ios/web setting.  Fixes [#159](https://github.com/jonbhanson/flutter_native_splash/issues/159).